requirements.txt
asgiref==3.3.4
class-registry==2.1.2
dj-database-url==0.5.0
Django==3.2.3
django-bootstrap-form==3.4
django-filter==2.4.0
django-heroku==0.3.1
django-materializecss-form==1.1.17
filters==1.3.2
gunicorn==20.1.0
humanize==3.8.0
Pillow==8.2.0
psycopg2==2.9.1
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
regex==2021.9.30
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
